# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  api.ProfileSettings:
    properties:
      api_active:
        type: boolean
      api_key:
        type: string
      auto_import_directory:
        type: string
      language:
        type: string
      prefer_full_date:
        type: boolean
      preferred_units:
        $ref: '#/definitions/database.UserPreferredUnits'
      socials_disabled:
        type: boolean
      theme:
        type: string
      timezone:
        type: string
      totals_show:
        type: string
    type: object
  api.RecordResponse:
    properties:
      date:
        type: string
      value:
        type: number
      workout_id:
        type: integer
    type: object
  api.Response-any:
    properties:
      errors:
        items:
          type: string
        type: array
      results: {}
    type: object
...
  api.Response-api_TotalsResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        $ref: '#/definitions/api.TotalsResponse'
    type: object
  api.Response-api_UserProfileResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        $ref: '#/definitions/api.UserProfileResponse'
    type: object
...
  api.Response-array_api_WorkoutRecordResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        items:
          $ref: '#/definitions/api.WorkoutRecordResponse'
        type: array
    type: object
...
  api.TotalsResponse:
    properties:
      distance:
        type: number
      down:
        type: number
      duration:
        description: Duration in seconds
        type: integer
      up:
        type: number
      workouts:
        type: integer
    type: object
  api.UserProfileResponse:
    properties:
      active:
        type: boolean
      admin:
        type: boolean
      birthdate:
        type: string
      created_at:
        type: string
      id:
        type: integer
      language:
        type: string
      last_version:
        type: string
      name:
        type: string
      prefer_full_date:
        type: boolean
      preferred_units:
        $ref: '#/definitions/database.UserPreferredUnits'
      profile:
        $ref: '#/definitions/api.ProfileSettings'
      socials_disabled:
        type: boolean
      theme:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  api.WorkoutRecordResponse:
    properties:
      active:
        type: boolean
      average_speed:
        $ref: '#/definitions/api.RecordResponse'
      average_speed_no_pause:
        $ref: '#/definitions/api.RecordResponse'
      distance:
        $ref: '#/definitions/api.RecordResponse'
      duration:
        $ref: '#/definitions/api.RecordResponse'
      max_speed:
        $ref: '#/definitions/api.RecordResponse'
      total_up:
        $ref: '#/definitions/api.RecordResponse'
      workout_type:
        type: string
    type: object
...
  database.UserPreferredUnits:
    properties:
      distance:
        description: The user's preferred distance unit
        type: string
      elevation:
        description: The user's preferred elevation unit
        type: string
      height:
        description: The user's preferred height unit
        type: string
      speed:
        description: The user's preferred speed unit
        type: string
      weight:
        description: The user's preferred weight unit
        type: string
    type: object
...
info:
  contact: {}
  description: Workout Tracker HTTP API (v2).
  title: Workout Tracker API
  version: "2.0"
paths:
...
  /admin/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Delete user (admin)
      tags:
      - admin
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Get user (admin)
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Update user (admin)
      tags:
      - admin
...
  /records:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-array_api_WorkoutRecordResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Get workout records
      tags:
      - user
...
  /totals:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_TotalsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Get workout totals
      tags:
      - user
  /whoami:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Get current user profile
      tags:
      - user
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  ApiKeyQuery:
    in: query
    name: api-key
    type: apiKey
  CookieAuth:
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"